Refactor Settings dialogs into separate component
Since the dialog is essentially reused for all of the different unit
types, put it into its own component.

Have the inputted model items be an array of two elements: the first
value is the actual value the settings should be set to, and the
second is the string to display to the user.